+++ citeproc/citeproc-node-names.lua
@@ -69,7 +69,7 @@ function Names:from_node(node)
         child.after_name = true
       end
     else
-      util.warning(string.format('Unkown element "{}".', element_name))
+      util.warning(string.format('Unknown element "%s".', element_name))
     end
   end
   o:get_delimiter_attribute(node)